% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotellipse.R
\name{plotellipse}
\alias{plotellipse}
\title{Plots an ellipse}
\usage{
plotellipse(ellipse)
}
\arguments{
\item{ellipse}{Gives the parameters of the ellipse - output from the ellipsoidhull functon in cluster.}
}
\value{
Nothing - just plots the ellipse.
}
\description{
Plots a minimum ellipse around a set of data
}
\details{
Routine adapted from a suggestion made on CrossValidated:  http://stats.stackexchange.com/questions/9898/how-to-plot-an-ellipse-from-eigenvalues-and-eigenvectors-in-r
}
\references{
Maechler, M., Rousseeuw, P., Struyf, A., Hubert, M., Hornik, K.(2013).
cluster: Cluster Analysis Basics and Extensions. R package version 1.14.4.
}

