% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait.r
\docType{data}
\name{trait}
\alias{trait}
\title{Sample data for use in convevol examples and tutorial}
\format{
I'll need to look into this
}
\description{
It's a data matrix with 6 variables measured across 100 tips of a phylogeny ("phy" - also included)
}
